requirements.txt
PyYAML~=6.0.2
Flask~=3.1.1
flask-swagger-ui~=5.21.0